Change sweeper to list all workspaces
The ephemeral workspaces we make during tests have random values.
Because sweepers are run in a separate command from the acceptance
tests, there's currently no way to persist those unique values between
commands so the sweeper can delete those specific workspaces.

Instead, as a first iteration, this change lists all workspaces in an
account and deletes any that match the acceptance testing prefix we use.
This is designed to run either by hand or by CI at a given interval,
such as overnight, when other acceptance tests are not running.

+++ internal/provider/resources/sweeper_test.go
@@ -12,38 +12,55 @@ import (
 	"github.com/prefecthq/terraform-provider-prefect/internal/testutils"
 )
 
+// TestMain adds sweeper functionality to the "go test" command.
 func TestMain(m *testing.M) {
 	resource.TestMain(m)
 }
 
-func addWorkspaceSweeper(testSuffix string, workspaces []string) {
-	sweeperName := fmt.Sprintf("workspaces_%s", testSuffix)
+//nolint:gochecknoinits // required per Terraform testing guide
+func init() {
+	addWorkspaceSweeper()
+}
 
-	resource.AddTestSweepers(sweeperName, &resource.Sweeper{
-		Name: sweeperName,
+// addWorkspaceSweeper adds a sweeper that deletes any workspaces that match
+// the prefix we use for ephemeral workspaces in acceptance tests.
+//
+// This is designed to run at a given interval when other acceptance tests are
+// not likely running.
+func addWorkspaceSweeper() {
+	resource.AddTestSweepers("workspaces", &resource.Sweeper{
+		Name: "workspaces",
 		F: func(_ string) error {
-			ctx := context.Background()
-			client, _ := testutils.NewTestClient()
+			client, err := testutils.NewTestClient()
+			if err != nil {
+				return fmt.Errorf("unable to get prefect client: %w", err)
+			}
 
 			// NOTE: the accountID is inherited by the one set in the test environment
-			workspacesClient, _ := client.Workspaces(uuid.Nil)
-
-			workspaces, err := workspacesClient.List(ctx, workspaces)
+			workspacesClient, err := client.Workspaces(uuid.Nil)
 			if err != nil {
-				log.Printf("unable to list workspaces: %v", err)
+				return fmt.Errorf("unable to get workspaces client: %w", err)
+			}
+
+			workspaces, err := workspacesClient.List(context.Background(), []string{})
+			if err != nil {
+				return fmt.Errorf("unable to list workspaces: %w", err)
+			}
+
+			if len(workspaces) == 0 {
+				return fmt.Errorf("no workspaces found for this account")
 			}
 
 			for _, workspace := range workspaces {
-				log.Printf("found workspace: %s", workspace.Name)
 				if strings.HasPrefix(workspace.Name, testutils.TestAccPrefix) {
-					log.Printf("found matching workspace %s, deleting...", workspace.Name)
-					err := workspacesClient.Delete(context.Background(), workspace.ID)
+					log.Printf("found acceptance testing workspace %s, deleting...\n", workspace.Name)
 
+					err := workspacesClient.Delete(context.Background(), workspace.ID)
 					if err != nil {
-						log.Printf("Error destroying %s during sweep: %s", workspace.Name, err)
+						log.Printf("unable to delete workspaces %s during sweep: %s\n", workspace.Name, err)
 					}
 				} else {
-					log.Printf("workspace %s did not match prefix, skipping...", workspace.Name)
+					log.Printf("workspace %s does not match acceptance testing prefix, skipping...\n", workspace.Name)
 				}
 			}
